Move library manga filtering into hook
